app/requirements.txt
fastapi
pydantic
uvicorn
pymongo
python-dotenv
passlib[bcrypt]
pyjwt

